Moved GLE spectrum deconvolution code to tools (#384)
* Moved GLE spectrum deconvolution code to tools
