+++ ipi/utils/tools/__init__.py
@@ -1,3 +1,9 @@
 from .acf_xyz import compute_acf_xyz
+from .gle import gle_frequency_kernel, isra_deconvolute, get_gle_matrices
 
-__all__ = ["compute_acf_xyz"]
+__all__ = [
+    "compute_acf_xyz",
+    "gle_frequency_kernel",
+    "isra_deconvolute",
+    "get_gle_matrices",
+]
